Fix a race condition when dispatching responses.
The old code dispatched responses in a goroutine that closed on
`message_packet`. Lexical closures in Go are by-reference, so the
closed-on pointer could change out from under the scheduled posting
(depending on how the scheduler feels).

This fix makes a copy of the field to send, such that if the result is
sent back after the next loop iteration, the old desired copy is still
the one sent back.

This bug was highlighted by `TestMultiGoroutineSearch` spitting back
"Unexpected Message Result" errors.