Fix a race condition causing panic: send on closed channel
This is a really, really nasty one to reproduce (and may have actually
been fixed by the previous commit -- not sure). Effectively, the result
channels could be closed out from under the result sender. This patch
adds more locking (using a `sync.RWMutex`) which prevents
`Conn.closeAllChannels` from starting when there's still a goroutine
running that wants to send a message on the channels being closed.

I really am adverse to the semantics this patch uses to achieve this --
to work, it acquires a read lock *before* spawning a goroutine, then
unlocks *inside* the goroutine. The godocs for `sync.RWMutex` don't
explicitly disallow this, but it seems very gross to me.